pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "figwizz"
version = "0.1.0"
description = "Toolkit for programmatic figure design"
readme = "README.md"
requires-python = ">=3.7"
license = "GPL-3.0-only"
authors = [
    { name = "Colin Conwell", email = "colinconwell@gmail.com" },
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
]

dependencies = [
    "ipython>=8.0.1",
    "pillow>=9.0.0",
    "python-pptx>=0.6.21",
    "pybtex>=0.24.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.62.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0"
]

[project.optional-dependencies]
genai = [
    "litellm>=1.55.0",
]
pdf = [
    "PyMuPDF>=1.22.5",
]
svg = [
    "cairosvg>=2.7.0",
]
numpy = [
    "numpy>=1.21.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "isort>=5.0.0",
    "black>=23.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.20.0",
]

[project.scripts]
figwizz = "figwizz.workflows.cli_opts:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["figwizz*"]

[tool.pytest.ini_options]
testpaths = ["testing"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers --tb=short"
markers = [
    "requires_pymupdf: mark test as requiring PyMuPDF",
    "requires_litellm: mark test as requiring litellm",
    "requires_numpy: mark test as requiring numpy",
    "requires_cairosvg: mark test as requiring cairosvg",
    "requires_network: mark test as requiring network access",
]

[tool.black]
line-length = 100
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100